add pass find. pass insert bugfix

+++ tests/Pass.Test.ps1
@@ -12,8 +12,10 @@ function Resolve-Error ($ErrorRecord = $Error[0]) {
 
 Import-Module $PSScriptRoot/../Pass.psm1 -PassThru
 $env:PASSWORD_STORE_DIR = "$PSScriptRoot/password-store-test/"
-Remove-Item -Recurse $env:PASSWORD_STORE_DIR -Force
 
+if (Test-Path $env:PASSWORD_STORE_DIR) {
+    Remove-Item -Recurse $env:PASSWORD_STORE_DIR -Force
+}
 try { 
     Invoke-PassInit -GpgId "test@example.com"
     Invoke-PassGit init
@@ -23,11 +25,13 @@ try {
     Invoke-PassInsert "example.com/test@example.com" "1234"
     Test-Path "$env:PASSWORD_STORE_DIR/example.com/test@example.com.gpg"
     Invoke-PassShow "example.com/test@example.com"
+    Invoke-PassInsert "example.com/test@example.com" "1234"
 
     Invoke-PassGenerate "example2.com/test@example.com" 16
     Test-Path "$env:PASSWORD_STORE_DIR/example2.com/test@example.com.gpg"
+    Invoke-PassGenerate "example2.com/test@example.com" 16
     Invoke-PassShow "example2.com/test@example.com"
-
+    Invoke-PassFind "example"
 }
 catch {
     Resolve-Error $Error
